[lld] Pass random.randint stop parameter as int.
It's required by Python 3.12, from https://docs.python.org/3.12/library/random.html "Changed in version 3.12: Automatic conversion of non-integer types is no longer supported."

Reviewed By: #lld-macho, int3

Differential Revision: https://reviews.llvm.org/D145826

+++ lld/test/MachO/tools/generate-cfi-funcs.py
@@ -23,7 +23,7 @@ def print_function(name):
   global lsda_odds
   have_lsda = (random.random() < lsda_odds)
   frame_size = random.randint(4, 64) * 16
-  frame_offset = -random.randint(0, (frame_size/16 - 4)) * 16
+  frame_offset = -random.randint(0, int(frame_size/16 - 4)) * 16
   global func_size_low, func_size_high
   func_size = random.randint(func_size_low, func_size_high) * 0x10
   func_size_high += 1
